requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
django-ckeditor
pandas>=1.2.4
protobuf>=3.19.4
PyJWT>=2.3.0
requests>=2.22.0
pytest>=6.2.5
aiogram>2.0
matplotlib>=3.1.0
